Move DashboardService + CampContactService to Application (#568)
Both are pure composers (no DbContext, no repositories). Relocating
closes two §2b violations. DI registrations updated, no behavior change.
