+++ tests/Humans.Application.Tests/Services/CampContactServiceTests.cs
@@ -1,6 +1,6 @@
 using AwesomeAssertions;
 using Humans.Application.Interfaces;
-using Humans.Infrastructure.Services;
+using Humans.Application.Services.Camps;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging.Abstractions;
 using NSubstitute;
